requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
mysql-connector-python==8.0.32
python-dotenv==1.0.0
pydantic==1.10.7
python-multipart==0.0.6
